// src/redux/actions/types/callingListsActionTypes.ts
export enum ECallingListsActionTypes {
  LOADING = 'CALLINGLISTS/LOADING',
  LOADING_RECORDS = 'CALLINGLISTS/LOADING_RECORDS',
  LOADING_CALLING_LIST = 'CALLINGLISTS/LOADING_CALLING_LIST',
  SET = 'CALLINGLISTS/SET',
  ADD_RECORDS = 'CALLINGLISTS/ADD_RECORDS',
  ADD_RECORDS_SINGLE = 'CALLINGLISTS/ADD_RECORDS_SINGLE',
  ADD_ITEMS = 'CALLINGLISTS/ADD_ITEMS',
  ADD_ACTIVE = 'CALLINGLISTS/ADD_ACTIVE',
  TOGLE_EXPANDE = 'CALLINGLISTS/TOGLE_EXPANDE',
  TOGLE_LOADING = 'CALLINGLISTS/TOGLE_LOADING',
  SET_POPUP = 'CALLINGLISTS/SET_POPUP',
  SET_EDIT_POPUP = 'CALLINGLISTS/SET_EDIT_POPUP',
  SET_EDITAll_POPUP = 'CALLINGLISTS/SET_EDITAll_POPUP',
  ADD_FIELDS = 'CALLINGLISTS/ADD_FIELDS',
  TOGLE_CHECKED_RECCORD = 'CALLINGLISTS/TOGLE_CHECKED_RECCORD',
  TOGLE_CHECKED_ALL = 'CALLINGLISTS/TOGLE_CHECKED_ALL',
  SET_TOTAL_COUNT = 'CALLINGLISTS/SET_TOTAL_COUNT',
  CHENGE_PARAM = 'CALLINGLISTS/CHENGE_PARAM',
  UPDATE_COLORS = 'CALLINGLISTS/UPDATE_COLORS',
  USE_FILTER = 'CALLINGLISTS/USE_FILTER',
  SET_FILTER = 'CALLINGLISTS/SET_FILTER',
  CHANGE_ORDER_BY_CONDITION = 'CALLINGLISTS/CANGE_ORDER_BY',
  CHANGE_WHERE_CONDITION = 'CALLINGLISTS/CANGE_WHERE_CONDITION',
  SET_INCLUDED = 'CALLINGLISTS/SET_INCLUDED',
  SET_SHOW_CUSTOM_FIELDS = 'CALLINGLISTS/SET_SHOW_CUSTOM_FIELDS',
  SET_NEGATION = 'CALLINGLISTS/SET_NEGATION',
  SET_TABLE_INFO = 'CALLINGLISTS/SET_TABLE_INFO',
  SET_VIEWING_FILTER = 'CALLINGLISTS/SET_VIEWING_FILTER',
  SET_LISTS_TOTAL = 'CALLINGLISTS/SET_LISTS_TOTAL',
}